Commit the the two python scripts and one serialized file.
I have been coding in python before leaning about git and github. This
is the first python program that I have written which have no bugs in
it. I did my best to make this program be like a full program for
recording customers data. Currently I am learning sqlite and by the
time I am done, I will be writing a complex application than this and
I will make sure I commit at every point I get to to show that the code
I have written did not come from nowhere.

customer.dat is a pickle file holding the data of the customers.
person_and_customer.py is a file containing two classes - Person class
and Customer class. Customer class is a child class of Person class.
customer.py is a file containing main class that calls other classes
base on the user interactions with the application.

One thing that I want to mention which is in the customer.py is that I
have created functions to get first name and last name from the user
and then make the first letter of each name capital. I have done this
because while I was testing the code, I used a friend who is having a
compound name to test the code. His name is Omogbai-Francis. at the
beginning of writing the code, I was applying .capitalize() function on
the user input to make the first letter of the names capital. But I saw
a downside of using that, and that is, with th name Omogbai-Francis, it
will turn it into Omogbai-francis, leaving the second name in the
compound name after "-" in lower case. So I decided to create my own
function which will make the first letter of the input made by the user
capital if the user use a small letter.

Thank you for viewing this code.