added if __name__ == "__main__": main()
This is a function call that avoids the main function to be called
when this python file is imported in another python script, but the
functions created in this module can be used in that python script.
It is not important though, I just felt like putting it there, because
I learnt it is a best practice.